requirements.txt
cachetools==5.2.0
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
google-api-core==2.10.1
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-speech==2.15.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
